Data Type mismatch Fix (check on GPU CI) (#2085)
* Device mismatch

* Device management fixed

* removed extra type conversion

* Update ignite/metrics/gan/fid.py

* Fix for dtype in IS

* IS return type to float

* Mypy fix

+++ ignite/metrics/gan/utils.py
@@ -88,7 +88,7 @@ class _BaseInceptionMetric(Metric):
             inputs = inputs.to(self._device)
 
         with torch.no_grad():
-            outputs = self._feature_extractor(inputs).to(self._device)
+            outputs = self._feature_extractor(inputs).to(self._device, dtype=torch.float64)
         self._check_feature_shapes(outputs)
 
         return outputs
